Unify epoll backend implementations
